+++ crates/clinker-record/src/storage.rs
@@ -17,14 +17,14 @@ use crate::Value;
 /// `Value::String(Box<str>)`.
 pub trait RecordStorage: Send + Sync {
     /// Resolve a field by name from the record at the given index.
-    fn resolve_field(&self, index: u32, name: &str) -> Option<&Value>;
+    fn resolve_field(&self, index: u64, name: &str) -> Option<&Value>;
 
     /// Resolve a qualified field (source.field) from the record at the given index.
-    fn resolve_qualified(&self, index: u32, source: &str, field: &str) -> Option<&Value>;
+    fn resolve_qualified(&self, index: u64, source: &str, field: &str) -> Option<&Value>;
 
     /// List all available field names for the record at the given index.
-    fn available_fields(&self, index: u32) -> Vec<&str>;
+    fn available_fields(&self, index: u64) -> Vec<&str>;
 
     /// Total number of records in this storage.
-    fn record_count(&self) -> u32;
+    fn record_count(&self) -> u64;
 }
